requirements.txt
praw
requests
transformers
matplotlib
wordcloud
vaderSentiment
jinja2
python-dotenv
markdown